Refactor: Update Android Gradle plugin and dependencies
This commit updates the Android Gradle plugin version from 8.2.1 to 8.7.3.
It also updates the Gradle wrapper distribution URL to version 8.12.
The NDK version is updated to 27.0.12077973.
Java compatibility is updated to version 17.
The application version is updated to 1.1.1+11.

Several Flutter dependencies were removed and some were updated.
The `intl` package was added as a direct dependency.
The `change_app_package_name` package was moved to dev_dependencies.
Navigation logic in `SettingPageBody`, `ChangePassword`, and `Nav` was changed from `pushReplacement` to `push` for the `Login` route.
Locale settings were added to `gradle.properties`.
The `upload-keystore.jks` file was deleted.